requirements.txt

python-telegram-bot[job-queue]
python-dotenv
requests
aiohttp
nest_asyncio
